issues fixed in brag section and image section

+++ app/Http/Controllers/ImageSectionController.php
@@ -55,7 +55,7 @@ class ImageSectionController extends Controller
             // Storage::disk('public-storage')->delete($section->section_image);
             // $data['section_image'] = $request->file('section_image')->store('uploads/section_images', 'public-storage');
             // Delete the old image from storage
-            if($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg") {
+            if($section->section_image && ($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg")) {
                 Storage::disk('public-storage')->delete($section->section_image);
             }
             
@@ -85,7 +85,7 @@ class ImageSectionController extends Controller
         $section = ImageSection::findOrFail($sectionId);
         
         // Delete the image associated with the section
-        if($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg") {
+        if($section->section_image && ($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg")) {
             Storage::disk('public-storage')->delete($section->section_image);
         }
